Break out storage classes into separate files
