+++ src/scripts/services/sync/local-storage.js
@@ -1,0 +1,19 @@
+export function LocalStorage($q) {
+  'ngInject';
+
+  return {
+    get: function() {
+      return $q.resolve(JSON.parse(localStorage.getItem('DIM')));
+    },
+
+    set: function(value) {
+      // TODO: remove in favor of localStorage?
+      localStorage.setItem('DIM', JSON.stringify(value));
+      return $q.resolve(value);
+    },
+
+    // TODO: disable if indexedDB is on
+    enabled: true,
+    name: 'LocalStorage'
+  };
+}
